feat: Simplify highlighting logic, refactor code
- Shiki's createSingletonShorthands already handles creating a singleton
highlighter and loading themes and languages, so it's used to simplify
and streamline the hooks logic
- throttleHighlighting function moved to `utils.ts`
